requirements.txt
opencv-python~=4.7.0.72
Pillow~=9.5.0
numba~=0.57.1
numpy~=1.23.5